Unify service to a single haproxy service resource
For runit, we now use the "legacy" service resource. This now works
similar to the approach done by the nginx cookbook. The upside of this
is that we do not have to deal with the hard-coded force-reload command
for the reload action in runit which broke our seamless reload.

We can now also get rid of the helper methods to select the correct
resource to notify as we can now always notify service[haproxy].