Run LWRP providers with use_inline_resources
This is necessary as the previous reload performed directly inside the
provider is performed directly after the LWRP provider exited and is NOT
actually delayed to the end of the chef run. This would result in way
too many reloads if multiple files have changed.

Users of the LWRP resources now have to manually reload the service,
e.g.:

```ruby
haproxy_template "foo" do
  notifies :reload, "service[haproxy]"
end
```